+++ jarboot-server/src/main/java/io/github/majianzheng/jarboot/service/impl/UpgradeServiceImpl.java
@@ -3,16 +3,19 @@ package io.github.majianzheng.jarboot.service.impl;
 import io.github.majianzheng.jarboot.api.constant.CommonConst;
 import io.github.majianzheng.jarboot.api.pojo.ServerRuntimeInfo;
 import io.github.majianzheng.jarboot.base.AgentManager;
+import io.github.majianzheng.jarboot.cluster.ClusterClientManager;
 import io.github.majianzheng.jarboot.common.CacheDirHelper;
 import io.github.majianzheng.jarboot.common.JarbootException;
 import io.github.majianzheng.jarboot.common.notify.FrontEndNotifyEventType;
 import io.github.majianzheng.jarboot.common.pojo.UpgradeProgress;
 import io.github.majianzheng.jarboot.common.utils.*;
+import io.github.majianzheng.jarboot.service.UpgradeStoreFileCallback;
 import io.github.majianzheng.jarboot.service.ServerRuntimeService;
 import io.github.majianzheng.jarboot.service.UpgradeService;
 import io.github.majianzheng.jarboot.utils.CommonUtils;
 import io.github.majianzheng.jarboot.utils.MessageUtils;
 import io.github.majianzheng.jarboot.utils.SettingUtils;
+import io.github.majianzheng.jarboot.utils.TaskUtils;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.commons.io.FileUtils;
 import org.springframework.stereotype.Service;
@@ -20,12 +23,14 @@ import org.springframework.util.CollectionUtils;
 
 import javax.annotation.Resource;
 import java.io.File;
+import java.io.IOException;
 import java.io.InputStream;
 import java.io.OutputStream;
 import java.util.Arrays;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicBoolean;
 
 /**
@@ -38,6 +43,11 @@ public class UpgradeServiceImpl implements UpgradeService {
     private final AtomicBoolean upgrading = new AtomicBoolean(false);
     @Resource
     private ServerRuntimeService serverRuntimeService;
+
+    @Override
+    public boolean isUpgrading() {
+        return upgrading.get();
+    }
 
     @Override
     public void upgrade(String url) {
@@ -61,7 +71,7 @@ public class UpgradeServiceImpl implements UpgradeService {
     }
 
     @Override
-    public void upgrade(String file, InputStream is) {
+    public void upgrade(String file, InputStream is, UpgradeStoreFileCallback callback) {
         if (!upgrading.compareAndSet(false, true)) {
             throw new JarbootException("正在升级中，请稍后...");
         }
@@ -71,6 +81,9 @@ public class UpgradeServiceImpl implements UpgradeService {
         }
         try {
             FileUtils.copyInputStreamToFile(is, instFile);
+            if (callback != null) {
+                callback.run(instFile);
+            }
         } catch (Exception e) {
             upgrading.set(false);
             throw new JarbootException(e);
@@ -111,16 +124,27 @@ public class UpgradeServiceImpl implements UpgradeService {
             checkDir(FileUtils.getFile(distDir, "plugins"));
             checkDir(FileUtils.getFile(distDir, "workspace"));
             updateProgress("开始升级...", CommonConst.STEP_START_UPGRADE);
-            startUpgradeScript(distDir);
+            backgroundStart(distDir);
         } catch (Exception e) {
+            upgrading.set(false);
             log.error("升级失败: {}", e.getMessage(), e);
             updateProgress(e.getMessage(), CommonConst.STEP_UPGRADE_FAILED);
-        } finally {
-            upgrading.set(false);
         }
     }
 
-    private void startUpgradeScript(File distDir) throws Exception {
+    private void backgroundStart(File distDir) {
+        TaskUtils.getTaskExecutor().schedule(() -> {
+            try {
+                startUpgradeScript(distDir);
+            } catch (Exception e) {
+                upgrading.compareAndSet(true, false);
+                log.error("升级失败", e);
+                updateProgress("升级失败" + e.getMessage(), CommonConst.STEP_UPGRADE_FAILED);
+            }
+        }, 3, TimeUnit.SECONDS);
+    }
+
+    private void startUpgradeScript(File distDir) throws IOException {
         String home = SettingUtils.getHomePath();
         File workDir;
         List<String> cmd;
@@ -132,8 +156,9 @@ public class UpgradeServiceImpl implements UpgradeService {
         } else {
             workDir = FileUtils.getFile(distDir, "bin");
             String command = String.format(
-                    "nohup bash upgrade.sh -d '%s' -y >/dev/null 2>&1 &",
-                    home.replace("'", "'\"'\"'") // 转义单引号
+                    "nohup bash upgrade.sh -d '%s' -y >\"%s/logs/upgrade.log\" 2>&1 &",
+                    home.replace("'", "'\"'\"'"), // 转义单引号
+                    home
             );
             cmd = Arrays.asList("bash", "-c", command);
         }
@@ -210,6 +235,7 @@ public class UpgradeServiceImpl implements UpgradeService {
         UpgradeProgress progress = new UpgradeProgress();
         progress.setAction(action);
         progress.setMsg(msg);
+        progress.setHost(ClusterClientManager.getInstance().getSelfHost());
         MessageUtils.globalEvent(JsonUtils.toJsonString(progress), FrontEndNotifyEventType.UPGRADE_PROGRESS);
     }