+++ jarboot-server/src/main/java/io/github/majianzheng/jarboot/task/TaskJob.java
@@ -1,6 +1,7 @@
 package io.github.majianzheng.jarboot.task;
 
 import io.github.majianzheng.jarboot.api.constant.CommonConst;
+import io.github.majianzheng.jarboot.api.constant.SettingPropConst;
 import io.github.majianzheng.jarboot.api.pojo.ServiceSetting;
 import io.github.majianzheng.jarboot.base.AgentManager;
 import io.github.majianzheng.jarboot.cluster.ClusterClientManager;
@@ -39,13 +40,13 @@ public class TaskJob extends QuartzJobBean {
     }
 
     private void startTask(String sid, String userDir, String name) {
-        if (AgentManager.getInstance().exist(sid)) {
+        MessageUtils.console(sid, "定时任务触发，开始执行...");
+        ServiceSetting setting = PropertyFileUtils.getServiceSetting(userDir, name);
+        if (SettingPropConst.SCHEDULE_CRON.equals(setting.getScheduleType()) && AgentManager.getInstance().exist(sid)) {
             MessageUtils.info(name + "正在运行中或启动中，定时任务跳过！");
             MessageUtils.console(sid, "正在运行中或启动中，无需再次执行！");
             return;
         }
-        MessageUtils.console(sid, "定时任务触发，开始执行...");
-        ServiceSetting setting = PropertyFileUtils.getServiceSetting(userDir, name);
         try {
             if (null != STARTING_MAP.putIfAbsent(sid, setting)) {
                 MessageUtils.info(name + "正在启动中，定时任务跳过！");
@@ -54,6 +55,9 @@ public class TaskJob extends QuartzJobBean {
             }
             //记录开始时间
             long startTime = System.currentTimeMillis();
+            if (SettingPropConst.RESTART_CRON.equals(setting.getScheduleType())) {
+                TaskUtils.killService(sid);
+            }
             TaskUtils.startService(setting);
             double costTime = (System.currentTimeMillis() - startTime)/1000.0f;
             String msg = String.format("定时任务\033[96;1m%s\033[0m 启动耗时 \033[91;1m%.3f\033[0m second.\033[5m✨\033[0m", name, costTime);